cors/CORS_013.htm

<!DOCTYPE html>
<html>
    <head>
        <title>Verify a withCredential XHR Request fails if Access-Control-Allow-Credentials = false</title>
        <meta http-equiv="content-type" content="text/html; charset=UTF-8" />
        <meta description="Verify a withCredential XHR Request fails if Access-Control-Allow-Credentials = false" />
        <link rel="author" title="Microsoft" href="http://www.microsoft.com/" />
        <link rel="help" href="http://dvcs.w3.org/hg/cors/raw-file/tip/Overview.html#resource-sharing-check-0"/>
        <script src="http://w3c-test.org/resources/testharness.js"></script>
        <script src="http://w3c-test.org/resources/testharnessreport.js"></script>
        <script src="SetCORSUrls.js"></script>
    </head>
    <body>
        <div id=log></div>
        <script type="text/javascript">
            var t = async_test("Verify a withCredential XHR Request fails if Access-Control-Allow-Credentials = false");
            var xhr = new XMLHttpRequest();
            if('withCredentials' in xhr)
            {
                var url = CORS_SERVER_URL + "?TestID=13"
                xhr.open("GET", url, true);
                xhr.withCredentials = true;
                xhr.onreadystatechange = function(e)
                {
                    if (xhr.readyState == 4)
                    {
                        // The server responds back with the header 'Access-Control-Allow-Credentials: false'
                        // Because the credentials flag is true and the Access-Control-Allow-Credentials
                        // header value is not a case-sensitive match for "true", the request should fail
                        t.step(function(){assert_true(xhr.status === 0 && xhr.responseText === "")});
                        t.done();
                    }
                }
                xhr.send();
            }
            else
            {
                t.step(function(){assert_unreached("[Cross-Origin XMLHttpRequests are not supported]")});
                t.done();
            }
        </script>
    </body>
</html>
